refactor(admin): rename index-utils.ts → index-name.ts; drop commons.js re-export
