+++ api/src/datasets/es/commons.js
@@ -732,5 +732,3 @@ export const prepareResultItem = (hit, dataset, query, flatten, publicBaseUrl = 
 }
 
 export { extractError, escapeFilter } from './operations.ts'
-
-export { parseIndexName } from './index-utils.ts'
